requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
gunicorn==20.1.0
python-dotenv==0.19.0
PyYAML==6.0 